[vpnc] Switch to new upstream and update to 0.5.3.r524
Switch to new upstream and update to 0.5.3.r524.
Remove old packaging cruft that was imported from Fedora, most of
these were related to the vpnc script that we don't use.
Switch anything besides vpnc programs such as default config, which
was only used if vpnc was started outside of a connection manager, to
the documentation package.

Signed-off-by: Björn Bidar <bjorn.bidar@jolla.com>